Rename startup.js and restructure files
